# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис безопасной одноразовой передачи конфиденциальных файлов.",
        "title": "BorrowTime API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "BorrowTime Team",
            "email": "support@borrowtime.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/s/{token}": {
            "get": {
                "description": "Публичный endpoint для получения зашифрованного файла по одноразовому токену.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Скачать файл по токену",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Одноразовый токен доступа",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email получателя (если требуется политикой)",
                        "name": "email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Зашифрованный файл",
                        "schema": {
                            "type": "file"
                        },
                        "headers": {
                            "X-Encryption-Alg": {
                                "type": "string",
                                "description": "Алгоритм шифрования"
                            },
                            "X-Encryption-IV": {
                                "type": "string",
                                "description": "Initialization Vector (base64)"
                            },
                            "X-Encryption-Tag": {
                                "type": "string",
                                "description": "Auth tag (base64)"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transfers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Загружает зашифрованный файл и создаёт ссылку для получателя. Требует аутентификации.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Создать передачу файла",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Зашифрованный файл",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Срок действия ссылки (RFC3339)",
                        "name": "policy_expires_at",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Максимальное количество скачиваний (0 — без ограничений)",
                        "name": "policy_max_downloads",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Требовать аутентификацию получателя",
                        "name": "policy_require_auth",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Разрешённые email-адреса (можно передать несколько раз)",
                        "name": "policy_allowed_emails",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Алгоритм шифрования (AES-256-GCM)",
                        "name": "encryption_alg",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Initialization Vector (base64)",
                        "name": "encryption_iv",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Auth tag (base64)",
                        "name": "encryption_tag",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "transfer_id, share_url, access_token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transfers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает метаданные передачи. Доступно только владельцу.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Получить детали передачи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID передачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Досрочно блокирует ссылку. Доступно только владельцу передачи.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Отозвать доступ к передаче",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID передачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: revoked",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}